Fix current style context scope.
