+++ client/src/Components/Overview/index.jsx
@@ -8,7 +8,7 @@ import Selection from './Selection.jsx';
 
 import Socials from './Socials.jsx';
 import currentProducts from '../../Contexts/CurProdContext.js';
-import currentStyle from '../../Contexts/CurStyleContext.js';
+
 import styled from 'styled-components';
 
 const StyleContainer = styled.div`
@@ -37,40 +37,10 @@ const Product = styled.div`
 padding-top: 100px;
 `
 
-function Overview() {
+function Overview({styles, setCurStyle}) {
   const { currentProd } = useContext(currentProducts);
-  const [styles, setStyles] = useState([]);
-  const [curStyle, setCurStyle] = useState({});
-
-  let id = currentProd.id;
-
-  function getStyles() {
-    // console.log(id);
-    axios.get('products/:product_id/styles', { params: { product_id: id } })
-      .then((res) => {
-        setStyles(res.data.results);
-        defaultGrabber(res.data.results);})
-      // console.log(res.data))
-      // eslint-disable-next-line no-console
-      .catch((err) => console.error('err fetching styles', err));
-  }
-
-  function defaultGrabber(items) {
-    for (var i = 0; i < items.length; i++) {
-      if (items[i]['default?'] === true) {
-        setCurStyle(items[i]);
-        return;
-      }
-    }
-  }
-  // const [loaded, setLoad] = useState(false);
-
-  useEffect(() => {
-    getStyles();
-  }, [currentProd]);
 
   return (
-    <currentStyle.Provider value={{ curStyle, setCurStyle }}>
       <div id="overview">
         <Header><LogoCart /></Header>
         <Space />
@@ -96,7 +66,6 @@ function Overview() {
         <br />
         <div id="share-socials"><Socials /></div>
       </div>
-    </currentStyle.Provider>
   );
 }
